Restructure project for ICD implementation
