+++ libportability-gfx/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "portability-gfx"
+version = "0.1.0"
+authors = ["Dzmitry Malyshau <kvark@mozilla.com>"]
+
+[lib]
+name = "portability_gfx"
+
+[dependencies.gfx-hal]
+#path = "../gfx/src/hal"
+git = "https://github.com/kvark/gfx-rs"
+branch = "portable"
+
+[dependencies.gfx-backend-vulkan]
+#path = "../gfx/src/backend/vulkan"
+git = "https://github.com/kvark/gfx-rs"
+branch = "portable"
+features = ["portable"]
