+++ libportability-gfx/src/handle.rs
